Move read_flac_file to flac_encoder
