+++ tests/recipes/cad1/task1/test_evaluate.py
@@ -13,7 +13,6 @@ from recipes.cad1.task1.baseline.evaluate import (
     ResultsFile,
     _evaluate_song_listener,
     make_song_listener_list,
-    read_flac_signal,
     set_song_seed,
 )
 
@@ -187,39 +186,3 @@ def test_evaluate_song_listener(
             rel=pytest.rel_tolerance,
             abs=pytest.abs_tolerance,
         )
-
-
-def test_read_flac_signal(tmp_path):
-    # Create a test FLAC file
-    np.random.seed(2023)
-
-    filename = tmp_path / "test.flac"
-
-    sample_rate = 16000
-    signal = np.random.rand(1600)
-    scale_factor = np.max(np.abs(signal))
-
-    signal_scaled = signal / scale_factor
-    signal_scaled = signal_scaled * 32768.0
-    signal_scaled = np.clip(signal_scaled, -32768.0, 32767.0)
-    signal_scaled = signal_scaled.astype(np.dtype("int16"))
-
-    flac_encoder = FlacEncoder()
-    flac_encoder.encode(signal_scaled, sample_rate, filename)
-
-    # Create a test scale factor file
-    scale_filename = tmp_path / "test.txt"
-    with open(scale_filename, "w", encoding="utf-8") as fp:
-        fp.write(str(scale_factor))
-
-    # Call the function and check the output
-    signal_out, sample_rate_out = read_flac_signal(filename)
-
-    # As a result of the quantization, the signal is not exactly the same
-    # after encoding and decoding, so I'm changing the tolerance
-    # for this test
-    # np.sum(signal_out) = 2190.4271092907347
-    # np.sum(signal) = 2190.494140495932
-
-    assert np.sum(signal_out) == pytest.approx(np.sum(signal), rel=1e-4, abs=1e-4)
-    assert sample_rate_out == sample_rate
